Fix timeline search and print
Change the print timeline code so that issues appear under the record
they're apart of instead of above it. This was a silly bug that made it
hard to read time line issues.

Changed the search timeline code so that it's possible to filter based
on the task name. Previously we could only filter based on the task name
which is user defined.