Cannot access wwwroot files in VS Code
There are two scenarios here:

Use `dotnet run` to launch the application. This will work smoothly and
everything will go exactly as you expect it to go.

Use `launch` from VS code. At this point all of the resources under
wwwroot will get a 404 response code. I cannot figure this out for the
life of me.

In case it matters: the colorization of the text output from asp.net
core dissapears when run in the VS code mode.