CWD controls wwwroot in development
When I copied by launch.json entry from my development process I did not
update the `cwd` entry. That mean that `cwd` was pointing to the root of
my solution not to my asp.net project. Hence all of the static content
was mapped from the wrong directory and issued a 404.

Would've been nice for asp.net razor to issue an error when it failed to
find wwwroot. The lack of this though lead to a lot of hair pulling on
my part.