Basic auto-triage support added
Added a really basic tool that will automate the idea of filing bugs
based on predefined queries. It's pretty basic right now but will be
expanding it shortly.

Will be really nice when I can use it to automatically keep bugs up to
date with the issues that are found.