Initial refactor of triage
After consideration decided to make triage a separate tool after all.
The most likely destination for this tool is automation and cloud
services. Hence I will need to factor out all of this into the shared
library layer anyways. Making this a separate app is a good forcing
function here