Remove query pretty string
Having a pretty query string was nice but it was distracting from the
job of making the site functional. Making it scalable to all of the
pages basically required me to rewrite most of the search infrastructure
because it's all `Mono.Options` based.

Once the site is functional and proven I may go back and do this. For
now though I'm going to keep functioning on making the site useful on
top of the existing infrastructure that I've created.