Push GitHubClient through the program
This is needed for a lot of the queries that I'm working on. Bit the
bullet and pushed it through all of the core services.

At the moment the code still allows for anonymous use of `GitHubClient`
but that is just a transition method. Once I can push auth through the
rest of the pipeline I can delete all of the anonymous access code.