Display merged pull requests
This has the ability to display merged pull request rates for a given
repository and build combination.

It is _really_ slow right now because it must use both the AzDO and
GitHub REST APIs. Need to add this information into the DB and query it
directly from there instead of this approach.