Remove IAzureClient
The purpose of `IAzureClient` was to cache queries to azure devops. The
idea of caching was correct but the layering was wrong. What to cache
and where to cache is better done as a layer on top of `DevOpsServer`
not a layer below it. Makes it more declarative and allows for better
flexibility in the implementation.