Moved Mono.Options to runfo only
The `Mono.Options` library was a means to an end when I needed `string`
based queries in the website. It was always a bit of an akward fit
because it didn't meet the general style of web based queries.

Now that I have a full DB model for data and it has a completely
separate `string` based query model it's time to move `Mono.Options`
back to the command line tool.