Basic test tracking model
This defines the core model for tracking issues. It has the core
information of what is being tracked, what definitions should it be
restricted to, is it active, etc ...

These types will need to evolve as we expand out the tracking and
information stored there. Imagine there will be lots of optional columns
defined that are non-null when doing certain types of tracking. But that
can evolve as we add features.