Optimize query
The query used in the tracking util to find a particular
`ModelBuildAttempt` was showing up as the most expensive query in our
profiling. Changed it to use the primary key of the `ModelBuild` table
rather than a combo of three properties. That shoud make it much more
efficient.

Found some other places doing inefficient lookups and fixed those as
well. Took a bit of time to make our Find / Get patterns here more
consistent.