Support tracking query for test results
This updates our tracking query support for helix test results.

This is basically a rework of the tracking infra for Helix. It's now
much more general and has the same flow as the other tracking types.
This does mean I need to invalidate all of the exist tracking issues
that were Helix specific. That's okay though because they didn't
actually work.

closes #64