Remove DateTimeOffset usage
The date type used in the actual column storage is different than the
date time used in the filters. That was introducing extra work in all of
the queries to convert the column type.

True this was an cheap operation but it's totally unnecessary work here.
Using the same date type throughout is simple to do here.