Fix Helix Logs Search
The Helix Logs Search started timing out in the last week. The
underlying reason is the page was trying to filter `ModelTestResults` on
the property `IsHelixTestResult`. That is not an indexed property and
making it one is expensive table size wise. It's also unclear that it
would actually help if it were created because we have to do so much
additional filtering on the results.

As a quick fix I decided to have the helix log search essentially
iterate build by build. That is easy to get the perf right on even
though it requires a lot more manual searching. Need to think a bit more
abuot this but for now it will do.

Further during profiling I noticed that all of my build predicates were
not hitting the indexes either. That is because they were filtering by
definition then start time while the index is in the reverse order. That
was a simple fix.