Fix Azure Function Build
On a machine with only .NET 5 installed the azure function build won't
complete. It errors with the following:

```
It was not possible to find any compatible framework version
[/home/runner/work/runfo/runfo/DevOps.Functions/DevOps.Functions.csproj]
/home/runner/.nuget/packages/microsoft.net.sdk.functions/3.0.6/build/Microsoft.NET.Sdk.Functions.Build.targets(41,5):
error : The framework 'Microsoft.NETCore.App', version '3.0.0' was not
found.
[/home/runner/work/runfo/runfo/DevOps.Functions/DevOps.Functions.csproj]
/home/runner/.nuget/packages/microsoft.net.sdk.functions/3.0.6/build/Microsoft.NET.Sdk.Functions.Build.targets(41,5):
error :   - The following frameworks were found:
[/home/runner/work/runfo/runfo/DevOps.Functions/DevOps.Functions.csproj]
```
My theory is that they are running a `netcoreapp3.0` tool during build
that isn't setup for roll forward hence the error. Installing two SDKs
to see if that fixes it.