Timelines perf issues (#87)
The first part of this change is updating the `ModelBuilds` to
`ModelTimelineIssues` index to have a number of included proerties.
These properties are the core of our predicates when filtering on
issues. Now that they are included properties this _significantly
reduces the number of row reads we must do as the properties are right
there to be scanned.

The second part of the change is to use `.Date` in all of our date range
queries. This _significantly_ increases the quality of some of our
queries, particularly the model timeline issues. The hard part is though
that this only reproduces when executing from EF. When I execute the
exact same queries in SQL I get great performance and it hits the
indexes that I would expect. I cannot figure this out and have ruled out
the obvious answers. For now logging the win and will keep researching
for better answers.