Changed tracking util to hit index
The tracking util needed to hit the correct indexes. The query should
really just be looking at the issues / tests associated with the current
attempt and filtering those. There is no need to go by started date
first as that is irrelevant at this phase. The only part that matters is
whether or not the attempt matches the other criteria.