Figured out the model delete issues
The core problem is that I wasn't respecting the intended lifetime for a
`DbContext` instance. The intent is to keep them alive for a single
operation yet I was keeping them alive for multiple operations. That was
causing the `ChangeTracker` to get confused about the state of the
objects and preventing delete operations from working.

Changed the tests to allow for more single operation semantics and it's
all working now.