Make deleting old builds more incremental
Have a situation where deleting old builds is failing because we have
too many test failures. That causes the function to timeout trying to
delete the build. Adjust the function code to start trimming down the
build in this case so that we at least make progress on every invocation
of the function