Enable bypassing Azure Key Vault
For whatever reason permissions to Azure Key Vault aren't working for me
so I plumbed through an alternative to do local development on runfo.
See Documentation/DevWithoutKeyVault for details on the workflow.

Code-wise the primary changes are checking the USE_KEYVAULT environment
variable and some refactoring of the GitHubClientFactory to support using
the logged in user's OAuth token rather than the Runfo GH app official
identity. I also added the IConfiguration.GetNonNull() extension method and
used it in various places to proactively generate errors when the environment
has not provided configuration settings the app is relying on.