Fix Microsoft.Extensions.Configuration deployment error (#183)
The function app was failing with the error

> Could not load file or assembly 'Microsoft.Extensions.Configuration, Version=7.0.0.0, Culture=neutral, PublicKeyToken=adb9793829ddae60'

This is true even though that DLL at that version was deployed with the function. The reason is that the function runtime is on .NET 6.0 and it loads several dependencies before it loads the function code. Anything loaded by the function runtime can't be overridden by the function app. In this case MS.E.Configuration was one of those extensions. More simply it was a mistake to use 7.0.0 here on a .NET 6 runtime, should've kept it at 6.0.1 from the start.